:recycle: 테이블 ID값 Int -> Long

+++ src/main/java/com/pricewagon/pricewagon/domain/history/repository/ProductHistoryCustomRepository.java
@@ -5,5 +5,5 @@ import java.util.Optional;
 import com.pricewagon.pricewagon.domain.history.entity.ProductHistory;
 
 public interface ProductHistoryCustomRepository {
-	Optional<ProductHistory> findFirstByProductIdAndPriceNot(Integer productId, Integer currentPrice);
+	Optional<ProductHistory> findFirstByProductIdAndPriceNot(Long productId, Integer currentPrice);
 }
